+++ attributes/nsswitch.rb
@@ -1,0 +1,1 @@
+default[:sys][:nsswitch] = String.new